[skip ci] Move response dto to admin package
