+++ response-dto/src/main/java/org/ehrbase/response/openehr/admin/AdminStatusResponseData.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ehrbase.response.openehr;
+package org.ehrbase.response.openehr.admin;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
